refactored usePostPatchPut
